mid: 0000245 msg: Error al crear i modificar Agrupacions de Fets Vitals
git-svn-id: https://svn.code.sf.net/p/rolsac/code/rolsac-core/branches/rolsac-1.1@390 58ef2255-6f3f-491f-b93b-e266552ec6fa

+++ moduls/persistence/src/org/ibit/rol/sac/persistence/ejb/AgrupacionHVFacadeEJB.java
@@ -61,9 +61,11 @@ public abstract class AgrupacionHVFacadeEJB extends HibernateEJB {
         Session session = getSession();
         try {
         	
-        	for (HechoVitalAgrupacionHV hechoVitalAgrupacionHV : llistaFetsVitalsOld) {
-				session.delete(hechoVitalAgrupacionHV);
-			}
+            for (HechoVitalAgrupacionHV hechoVitalAgrupacionHV : llistaFetsVitalsOld) {
+                if (hechoVitalAgrupacionHV != null){
+                    session.delete(hechoVitalAgrupacionHV);   
+                }                    
+            }                   
         	
         	List<HechoVitalAgrupacionHV> listaHechoVitalAgrupacionHV = hechov.getHechosVitalesAgrupacionHV();
